Fix/bug empty link (#27)

+++ src/Observers/ItemObserver.php
@@ -70,7 +70,7 @@ class ItemObserver
                 }
             }
 
-            if (request()->get('link_type') === MenuItem::TYPE_EMPTY) {
+            if ((int) request()->get('link_type') === MenuItem::TYPE_EMPTY) {
                 $item->{menuItem::linkTypesAttributes()[MenuItem::TYPE_EMPTY]} = 1;
             } else {
                 $item->{menuItem::linkTypesAttributes()[MenuItem::TYPE_EMPTY]} = 0;
